website-book-chatbot/requirements.txt


fastapi==0.110.0
uvicorn[standard]==0.27.1



python-dotenv==1.0.1
pydantic==2.6.3



sqlalchemy==2.0.28
psycopg2-binary==2.9.9


langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-google-genai==0.0.11
google-generativeai==0.4.1